tests/baseline/protocol-1.2/http/tests/inspector/sources/debugger/async-callstack-network-initiator-expected.txt

CONSOLE MESSAGE: line 11: console.clear
Tests asynchronous call stacks printed in console for a Network.Initiator.

Set timer for test function.
async-callstack-network-initiator.html:11 Console was cleared
async-callstack-network-initiator.html:29 POST http://127.0.0.1:8000/failure/foo 404 (Not Found)

sendXHR @ async-callstack-network-initiator.html:29

timeout2 @ async-callstack-network-initiator.html:22

setTimeout (async)

timeout1 @ async-callstack-network-initiator.html:17

setTimeout (async)

testFunction @ async-callstack-network-initiator.html:12
